Error sidebar working
