+++ jsettlers.mapcreator/src/jsettlers/mapcreator/localization/labels_de.properties
@@ -30,8 +30,9 @@ action.quit          = Beenden
 action.redo          = Widerholen
 action.save          = Speichern
 action.save-as       = Speichern unter...
-action.show-errors   = Fehler
+action.show-errors   = Fehler anzeigen
 action.show-errors_n = %d Fehler
+action.show-tools    = Werkzeuge anzeigen
 action.statistic     = Kartenstatistiken anzeigen
 action.undo          = R\u00FCckg\u00E4ngig
 action.zoom-in       = Vergr\u00F6ssern
@@ -202,6 +203,9 @@ setstartpoint = Startpunkt setzen
 settings = Einstellungen
 
 shapes = Formen
+
+sidebar.errors = Fehler
+sidebar.tools  = Werkzeuge
 
 stackdescr = %d %s platzieren
