+++ jsettlers.mapcreator/src/jsettlers/mapcreator/localization/labels_en_US.properties
@@ -30,8 +30,9 @@ action.quit          = Quit editor
 action.redo          = Redo
 action.save          = Save
 action.save-as       = Save as...
-action.show-errors   = errors
+action.show-errors   = Show errors
 action.show-errors_n = %d errors
+action.show-tools    = Show tools
 action.statistic     = Display map Statistics
 action.undo          = Undo
 action.zoom-in       = Zoom in
@@ -200,6 +201,9 @@ setstartpoint = Set start point
 settings = settings
 
 shapes = shapes
+
+sidebar.errors = Errors
+sidebar.tools  = Tools
 
 stackdescr = place %d %s
