+++ jsettlers.mapcreator/src/jsettlers/mapcreator/main/window/menu.properties
@@ -9,7 +9,7 @@ menubar=file,edit,view,help
 
 menu.file=new,open,save,save-as,---,play,---,export-image,---,quit
 menu.edit=undo,redo,---,map-settings
-menu.view=zoom-in,zoom-out,zoom100,---,statistic,---,show-errors,goto-error
+menu.view=zoom-in,zoom-out,zoom100,---,statistic,---,show-tools,---,show-errors,goto-error
 menu.help=manual,about
 
 ## Toolbar
