+++ jsettlers.mapcreator/src/jsettlers/mapcreator/main/window/shortcut.properties
@@ -10,3 +10,7 @@ redo = control Y
 zoom100 = control 0
 
 play = F12
+
+show-tools = F6
+show-errors = F7
+goto-error = F8
